Network.AWS.CloudDirectory.AttachToIndex

Description

Attaches the specified object to the specified index.

Creating a Request

attachToIndex

Arguments

:: Text
atiDirectoryARN

-> ObjectReference
atiIndexReference

-> ObjectReference
atiTargetReference

-> AttachToIndex

Creates a value of AttachToIndex with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atiDirectoryARN - The Amazon Resource Name (ARN) of the directory where the object and index exist.

atiIndexReference - A reference to the index that you are attaching the object to.

atiTargetReference - A reference to the object that you are attaching to the index.

data AttachToIndex

See: attachToIndex smart constructor.

Request Lenses

atiDirectoryARN :: Lens' AttachToIndex Text

The Amazon Resource Name (ARN) of the directory where the object and index exist.

atiIndexReference :: Lens' AttachToIndex ObjectReference

A reference to the index that you are attaching the object to.

atiTargetReference :: Lens' AttachToIndex ObjectReference

A reference to the object that you are attaching to the index.

Destructuring the Response

attachToIndexResponse

Arguments

:: Int
atirsResponseStatus

-> AttachToIndexResponse

Creates a value of AttachToIndexResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atirsAttachedObjectIdentifier - The ObjectIdentifier of the object that was attached to the index.

atirsResponseStatus - -- | The response status code.

data AttachToIndexResponse

See: attachToIndexResponse smart constructor.

Response Lenses

atirsAttachedObjectIdentifier :: Lens' AttachToIndexResponse (Maybe Text)

The ObjectIdentifier of the object that was attached to the index.

atirsResponseStatus :: Lens' AttachToIndexResponse Int

- | The response status code.