Network.AWS.CloudDirectory.CreateDirectory

Description

Creates a Directory by copying the published schema into the directory. A directory cannot be created without a schema.

Creating a Request

createDirectory

Arguments

:: Text
cdName

-> Text
cdSchemaARN

-> CreateDirectory

Creates a value of CreateDirectory with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdName - The name of the Directory . Should be unique per account, per region.

cdSchemaARN - The Amazon Resource Name (ARN) of the published schema that will be copied into the data Directory . For more information, see arns .

data CreateDirectory

See: createDirectory smart constructor.

Request Lenses

cdName :: Lens' CreateDirectory Text

The name of the Directory . Should be unique per account, per region.

cdSchemaARN :: Lens' CreateDirectory Text

The Amazon Resource Name (ARN) of the published schema that will be copied into the data Directory . For more information, see arns .

Destructuring the Response

createDirectoryResponse

Arguments

:: Int
cdrsResponseStatus

-> Text
cdrsDirectoryARN

-> Text
cdrsName

-> Text
cdrsObjectIdentifier

-> Text
cdrsAppliedSchemaARN

-> CreateDirectoryResponse

Creates a value of CreateDirectoryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdrsResponseStatus - -- | The response status code.

cdrsDirectoryARN - The ARN that is associated with the Directory . For more information, see arns .

cdrsName - The name of the Directory .

cdrsObjectIdentifier - The root object node of the created directory.

cdrsAppliedSchemaARN - The ARN of the published schema in the Directory . Once a published schema is copied into the directory, it has its own ARN, which is referred to applied schema ARN. For more information, see arns .

data CreateDirectoryResponse

See: createDirectoryResponse smart constructor.

Response Lenses

cdrsResponseStatus :: Lens' CreateDirectoryResponse Int

- | The response status code.

cdrsDirectoryARN :: Lens' CreateDirectoryResponse Text

The ARN that is associated with the Directory . For more information, see arns .

cdrsName :: Lens' CreateDirectoryResponse Text

The name of the Directory .

cdrsObjectIdentifier :: Lens' CreateDirectoryResponse Text

The root object node of the created directory.

cdrsAppliedSchemaARN :: Lens' CreateDirectoryResponse Text

The ARN of the published schema in the Directory . Once a published schema is copied into the directory, it has its own ARN, which is referred to applied schema ARN. For more information, see arns .