Network.AWS.CloudDirectory.CreateIndex

Description

Creates an index object. See Indexing for more information.

Creating a Request

createIndex

Arguments

:: Text
ciDirectoryARN

-> Bool
ciIsUnique

-> CreateIndex

Creates a value of CreateIndex with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciParentReference - A reference to the parent object that contains the index object.

ciLinkName - The name of the link between the parent object and the index object.

ciDirectoryARN - The ARN of the directory where the index should be created.

ciOrderedIndexedAttributeList - Specifies the attributes that should be indexed on. Currently only a single attribute is supported.

ciIsUnique - Indicates whether the attribute that is being indexed has unique values or not.

data CreateIndex

See: createIndex smart constructor.

Request Lenses

ciParentReference :: Lens' CreateIndex (Maybe ObjectReference)

A reference to the parent object that contains the index object.

ciLinkName :: Lens' CreateIndex (Maybe Text)

The name of the link between the parent object and the index object.

ciDirectoryARN :: Lens' CreateIndex Text

The ARN of the directory where the index should be created.

ciOrderedIndexedAttributeList :: Lens' CreateIndex [AttributeKey]

Specifies the attributes that should be indexed on. Currently only a single attribute is supported.

ciIsUnique :: Lens' CreateIndex Bool

Indicates whether the attribute that is being indexed has unique values or not.

Destructuring the Response

createIndexResponse

Arguments

:: Int
cirsResponseStatus

-> CreateIndexResponse

Creates a value of CreateIndexResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cirsObjectIdentifier - The ObjectIdentifier of the index created by this operation.

cirsResponseStatus - -- | The response status code.

data CreateIndexResponse

See: createIndexResponse smart constructor.

Response Lenses

cirsObjectIdentifier :: Lens' CreateIndexResponse (Maybe Text)

The ObjectIdentifier of the index created by this operation.

cirsResponseStatus :: Lens' CreateIndexResponse Int

- | The response status code.