Network.AWS.CloudDirectory.CreateSchema

Description

Creates a new schema in a development state. A schema can exist in three phases:

Development: This is a mutable phase of the schema. All new schemas are in the development phase. Once the schema is finalized, it can be published.

Published: Published schemas are immutable and have a version associated with them.

Applied: Applied schemas are mutable in a way that allows you to add new schema facets. You can also add new, nonrequired attributes to existing schema facets. You can apply only published schemas to directories.

Creating a Request

createSchema

Arguments

:: Text
csName

-> CreateSchema

Creates a value of CreateSchema with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csName - The name that is associated with the schema. This is unique to each account and in each region.

data CreateSchema

See: createSchema smart constructor.

Request Lenses

csName :: Lens' CreateSchema Text

The name that is associated with the schema. This is unique to each account and in each region.

Destructuring the Response

createSchemaResponse

Arguments

:: Int
csrsResponseStatus

-> CreateSchemaResponse

Creates a value of CreateSchemaResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csrsSchemaARN - The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns .

csrsResponseStatus - -- | The response status code.

data CreateSchemaResponse

See: createSchemaResponse smart constructor.

Response Lenses

csrsSchemaARN :: Lens' CreateSchemaResponse (Maybe Text)

The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns .

csrsResponseStatus :: Lens' CreateSchemaResponse Int

- | The response status code.