Network.AWS.CloudDirectory.DeleteDirectory

Description

Deletes a directory. Only disabled directories can be deleted. A deleted directory cannot be undone. Exercise extreme caution when deleting directories.

Creating a Request

deleteDirectory

Arguments

:: Text
delDirectoryARN

-> DeleteDirectory

Creates a value of DeleteDirectory with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delDirectoryARN - The ARN of the directory to delete.

data DeleteDirectory

See: deleteDirectory smart constructor.

Request Lenses

delDirectoryARN :: Lens' DeleteDirectory Text

The ARN of the directory to delete.

Destructuring the Response

deleteDirectoryResponse

Arguments

:: Int
ddrsResponseStatus

-> Text
ddrsDirectoryARN

-> DeleteDirectoryResponse

Creates a value of DeleteDirectoryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddrsResponseStatus - -- | The response status code.

ddrsDirectoryARN - The ARN of the deleted directory.

data DeleteDirectoryResponse

See: deleteDirectoryResponse smart constructor.

Response Lenses

ddrsResponseStatus :: Lens' DeleteDirectoryResponse Int

- | The response status code.

ddrsDirectoryARN :: Lens' DeleteDirectoryResponse Text

The ARN of the deleted directory.