Network.AWS.CloudDirectory.DeleteFacet

Description

Deletes a given Facet . All attributes and Rule s that are associated with the facet will be deleted. Only development schema facets are allowed deletion.

Creating a Request

deleteFacet

Arguments

:: Text
dfSchemaARN

-> Text
dfName

-> DeleteFacet

Creates a value of DeleteFacet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfSchemaARN - The Amazon Resource Name (ARN) that is associated with the Facet . For more information, see arns .

dfName - The name of the facet to delete.

data DeleteFacet

See: deleteFacet smart constructor.

Request Lenses

dfSchemaARN :: Lens' DeleteFacet Text

The Amazon Resource Name (ARN) that is associated with the Facet . For more information, see arns .

dfName :: Lens' DeleteFacet Text

The name of the facet to delete.

Destructuring the Response

deleteFacetResponse

Arguments

:: Int
dfrsResponseStatus

-> DeleteFacetResponse

Creates a value of DeleteFacetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfrsResponseStatus - -- | The response status code.

data DeleteFacetResponse

See: deleteFacetResponse smart constructor.

Response Lenses

dfrsResponseStatus :: Lens' DeleteFacetResponse Int

- | The response status code.