Network.AWS.CloudDirectory.DeleteObject

Description

Deletes an object and its associated attributes. Only objects with no children and no parents can be deleted.

Creating a Request

deleteObject

Arguments

:: Text
doDirectoryARN

-> ObjectReference
doObjectReference

-> DeleteObject

Creates a value of DeleteObject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

doDirectoryARN - The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns .

doObjectReference - A reference that identifies the object.

data DeleteObject

See: deleteObject smart constructor.

Request Lenses

doDirectoryARN :: Lens' DeleteObject Text

The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns .

doObjectReference :: Lens' DeleteObject ObjectReference

A reference that identifies the object.

Destructuring the Response

deleteObjectResponse

Arguments

:: Int
dorsResponseStatus

-> DeleteObjectResponse

Creates a value of DeleteObjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dorsResponseStatus - -- | The response status code.

data DeleteObjectResponse

See: deleteObjectResponse smart constructor.

Response Lenses

dorsResponseStatus :: Lens' DeleteObjectResponse Int

- | The response status code.