Network.AWS.CloudDirectory.DetachObject

Description

Detaches a given object from the parent object. The object that is to be detached from the parent is specified by the link name.

Creating a Request

detachObject

Arguments

:: Text
detDirectoryARN

-> ObjectReference
detParentReference

-> Text
detLinkName

-> DetachObject

Creates a value of DetachObject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

detDirectoryARN - The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns .

detParentReference - The parent reference from which the object with the specified link name is detached.

detLinkName - The link name associated with the object that needs to be detached.

data DetachObject

See: detachObject smart constructor.

Request Lenses

detDirectoryARN :: Lens' DetachObject Text

The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns .

detParentReference :: Lens' DetachObject ObjectReference

The parent reference from which the object with the specified link name is detached.

detLinkName :: Lens' DetachObject Text

The link name associated with the object that needs to be detached.

Destructuring the Response

detachObjectResponse

Arguments

:: Int
detrsResponseStatus

-> DetachObjectResponse

Creates a value of DetachObjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

detrsDetachedObjectIdentifier - The ObjectIdentifier that was detached from the object.

detrsResponseStatus - -- | The response status code.

data DetachObjectResponse

See: detachObjectResponse smart constructor.

Response Lenses

detrsDetachedObjectIdentifier :: Lens' DetachObjectResponse (Maybe Text)

The ObjectIdentifier that was detached from the object.

detrsResponseStatus :: Lens' DetachObjectResponse Int

- | The response status code.