Network.AWS.CloudDirectory.DetachPolicy

Description

Detaches a policy from an object.

Creating a Request

detachPolicy

Arguments

:: Text
dpDirectoryARN

-> ObjectReference
dpPolicyReference

-> ObjectReference
dpObjectReference

-> DetachPolicy

Creates a value of DetachPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpDirectoryARN - The Amazon Resource Name (ARN) that is associated with the Directory where both objects reside. For more information, see arns .

dpPolicyReference - Reference that identifies the policy object.

dpObjectReference - Reference that identifies the object whose policy object will be detached.

data DetachPolicy

See: detachPolicy smart constructor.

Request Lenses

dpDirectoryARN :: Lens' DetachPolicy Text

The Amazon Resource Name (ARN) that is associated with the Directory where both objects reside. For more information, see arns .

dpPolicyReference :: Lens' DetachPolicy ObjectReference

Reference that identifies the policy object.

dpObjectReference :: Lens' DetachPolicy ObjectReference

Reference that identifies the object whose policy object will be detached.

Destructuring the Response

detachPolicyResponse

Arguments

:: Int
dprsResponseStatus

-> DetachPolicyResponse

Creates a value of DetachPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dprsResponseStatus - -- | The response status code.

data DetachPolicyResponse

See: detachPolicyResponse smart constructor.

Response Lenses

dprsResponseStatus :: Lens' DetachPolicyResponse Int

- | The response status code.