Network.AWS.CloudDirectory.GetObjectInformation

Description

Retrieves metadata about an object.

Creating a Request

getObjectInformation

Arguments

:: Text
goiDirectoryARN

-> ObjectReference
goiObjectReference

-> GetObjectInformation

Creates a value of GetObjectInformation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

goiConsistencyLevel - The consistency level at which to retrieve the object information.

goiDirectoryARN - The ARN of the directory being retrieved.

goiObjectReference - A reference to the object.

data GetObjectInformation

See: getObjectInformation smart constructor.

Request Lenses

goiConsistencyLevel :: Lens' GetObjectInformation (Maybe ConsistencyLevel)

The consistency level at which to retrieve the object information.

goiDirectoryARN :: Lens' GetObjectInformation Text

The ARN of the directory being retrieved.

goiObjectReference :: Lens' GetObjectInformation ObjectReference

A reference to the object.

Destructuring the Response

getObjectInformationResponse

Arguments

:: Int
goirsResponseStatus

-> GetObjectInformationResponse

Creates a value of GetObjectInformationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

goirsObjectIdentifier - The ObjectIdentifier of the specified object.

goirsSchemaFacets - The facets attached to the specified object.

goirsResponseStatus - -- | The response status code.

data GetObjectInformationResponse

See: getObjectInformationResponse smart constructor.

Response Lenses

goirsObjectIdentifier :: Lens' GetObjectInformationResponse (Maybe Text)

The ObjectIdentifier of the specified object.

goirsSchemaFacets :: Lens' GetObjectInformationResponse [SchemaFacet]

The facets attached to the specified object.

goirsResponseStatus :: Lens' GetObjectInformationResponse Int

- | The response status code.