Network.AWS.CloudDirectory.GetSchemaAsJSON

Description

Retrieves a JSON representation of the schema. See JSON Schema Format for more information.

Creating a Request

getSchemaAsJSON

Arguments

:: Text
gsajSchemaARN

-> GetSchemaAsJSON

Creates a value of GetSchemaAsJSON with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsajSchemaARN - The ARN of the schema to retrieve.

data GetSchemaAsJSON

See: getSchemaAsJSON smart constructor.

Request Lenses

gsajSchemaARN :: Lens' GetSchemaAsJSON Text

The ARN of the schema to retrieve.

Destructuring the Response

getSchemaAsJSONResponse

Arguments

:: Int
gsajrsResponseStatus

-> GetSchemaAsJSONResponse

Creates a value of GetSchemaAsJSONResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsajrsDocument - The JSON representation of the schema document.

gsajrsName - The name of the retrieved schema.

gsajrsResponseStatus - -- | The response status code.

data GetSchemaAsJSONResponse

See: getSchemaAsJSONResponse smart constructor.

Response Lenses

gsajrsDocument :: Lens' GetSchemaAsJSONResponse (Maybe Text)

The JSON representation of the schema document.

gsajrsName :: Lens' GetSchemaAsJSONResponse (Maybe Text)

The name of the retrieved schema.

gsajrsResponseStatus :: Lens' GetSchemaAsJSONResponse Int

- | The response status code.