Network.AWS.CloudDirectory.ListAttachedIndices

Description

Lists indices attached to an object.

Creating a Request

listAttachedIndices

Arguments

:: Text
laiDirectoryARN

-> ObjectReference
laiTargetReference

-> ListAttachedIndices

Creates a value of ListAttachedIndices with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laiConsistencyLevel - The consistency level to use for this operation.

laiNextToken - The pagination token.

laiMaxResults - The maximum number of results to retrieve.

laiDirectoryARN - The ARN of the directory.

laiTargetReference - A reference to the object that has indices attached.

data ListAttachedIndices

See: listAttachedIndices smart constructor.

Request Lenses

laiConsistencyLevel :: Lens' ListAttachedIndices (Maybe ConsistencyLevel)

The consistency level to use for this operation.

laiNextToken :: Lens' ListAttachedIndices (Maybe Text)

The pagination token.

laiMaxResults :: Lens' ListAttachedIndices (Maybe Natural)

The maximum number of results to retrieve.

laiDirectoryARN :: Lens' ListAttachedIndices Text

The ARN of the directory.

laiTargetReference :: Lens' ListAttachedIndices ObjectReference

A reference to the object that has indices attached.

Destructuring the Response

listAttachedIndicesResponse

Arguments

:: Int
lairsResponseStatus

-> ListAttachedIndicesResponse

Creates a value of ListAttachedIndicesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lairsIndexAttachments - The indices attached to the specified object.

lairsNextToken - The pagination token.

lairsResponseStatus - -- | The response status code.

data ListAttachedIndicesResponse

See: listAttachedIndicesResponse smart constructor.

Response Lenses

lairsIndexAttachments :: Lens' ListAttachedIndicesResponse [IndexAttachment]

The indices attached to the specified object.

lairsNextToken :: Lens' ListAttachedIndicesResponse (Maybe Text)

The pagination token.

lairsResponseStatus :: Lens' ListAttachedIndicesResponse Int

- | The response status code.