Network.AWS.CloudDirectory.ListFacetAttributes

Description

Retrieves attributes attached to the facet.

Creating a Request

listFacetAttributes

Arguments

:: Text
lfaSchemaARN

-> Text
lfaName

-> ListFacetAttributes

Creates a value of ListFacetAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfaNextToken - The pagination token.

lfaMaxResults - The maximum number of results to retrieve.

lfaSchemaARN - The ARN of the schema where the facet resides.

lfaName - The name of the facet whose attributes will be retrieved.

data ListFacetAttributes

See: listFacetAttributes smart constructor.

Request Lenses

lfaNextToken :: Lens' ListFacetAttributes (Maybe Text)

The pagination token.

lfaMaxResults :: Lens' ListFacetAttributes (Maybe Natural)

The maximum number of results to retrieve.

lfaSchemaARN :: Lens' ListFacetAttributes Text

The ARN of the schema where the facet resides.

lfaName :: Lens' ListFacetAttributes Text

The name of the facet whose attributes will be retrieved.

Destructuring the Response

listFacetAttributesResponse

Arguments

:: Int
lfarsResponseStatus

-> ListFacetAttributesResponse

Creates a value of ListFacetAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfarsNextToken - The pagination token.

lfarsAttributes - The attributes attached to the facet.

lfarsResponseStatus - -- | The response status code.

data ListFacetAttributesResponse

See: listFacetAttributesResponse smart constructor.

Response Lenses

lfarsNextToken :: Lens' ListFacetAttributesResponse (Maybe Text)

The pagination token.

lfarsAttributes :: Lens' ListFacetAttributesResponse [FacetAttribute]

The attributes attached to the facet.

lfarsResponseStatus :: Lens' ListFacetAttributesResponse Int

- | The response status code.