Network.AWS.CloudDirectory.ListFacetNames

Description

Retrieves the names of facets that exist in a schema.

Creating a Request

listFacetNames

Arguments

:: Text
lfnSchemaARN

-> ListFacetNames

Creates a value of ListFacetNames with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfnNextToken - The pagination token.

lfnMaxResults - The maximum number of results to retrieve.

lfnSchemaARN - The Amazon Resource Name (ARN) to retrieve facet names from.

data ListFacetNames

See: listFacetNames smart constructor.

Request Lenses

lfnNextToken :: Lens' ListFacetNames (Maybe Text)

The pagination token.

lfnMaxResults :: Lens' ListFacetNames (Maybe Natural)

The maximum number of results to retrieve.

lfnSchemaARN :: Lens' ListFacetNames Text

The Amazon Resource Name (ARN) to retrieve facet names from.

Destructuring the Response

listFacetNamesResponse

Arguments

:: Int
lfnrsResponseStatus

-> ListFacetNamesResponse

Creates a value of ListFacetNamesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfnrsNextToken - The pagination token.

lfnrsFacetNames - The names of facets that exist within the schema.

lfnrsResponseStatus - -- | The response status code.

data ListFacetNamesResponse

See: listFacetNamesResponse smart constructor.

Response Lenses

lfnrsNextToken :: Lens' ListFacetNamesResponse (Maybe Text)

The pagination token.

lfnrsFacetNames :: Lens' ListFacetNamesResponse [Text]

The names of facets that exist within the schema.

lfnrsResponseStatus :: Lens' ListFacetNamesResponse Int

- | The response status code.