Network.AWS.CloudDirectory.ListIndex

Description

Lists objects attached to the specified index.

Creating a Request

listIndex

Arguments

:: Text
liDirectoryARN

-> ObjectReference
liIndexReference

-> ListIndex

Creates a value of ListIndex with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

liRangesOnIndexedValues - Specifies the ranges of indexed values that you want to query.

liConsistencyLevel - The consistency level to execute the request at.

liNextToken - The pagination token.

liMaxResults - The maximum number of results to retrieve from the index.

liDirectoryARN - The ARN of the directory that the index exists in.

liIndexReference - The reference to the index to list.

data ListIndex

See: listIndex smart constructor.

Request Lenses

liRangesOnIndexedValues :: Lens' ListIndex [ObjectAttributeRange]

Specifies the ranges of indexed values that you want to query.

liConsistencyLevel :: Lens' ListIndex (Maybe ConsistencyLevel)

The consistency level to execute the request at.

liNextToken :: Lens' ListIndex (Maybe Text)

The pagination token.

liMaxResults :: Lens' ListIndex (Maybe Natural)

The maximum number of results to retrieve from the index.

liDirectoryARN :: Lens' ListIndex Text

The ARN of the directory that the index exists in.

liIndexReference :: Lens' ListIndex ObjectReference

The reference to the index to list.

Destructuring the Response

listIndexResponse

Arguments

:: Int
lirsResponseStatus

-> ListIndexResponse

Creates a value of ListIndexResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lirsIndexAttachments - The objects and indexed values attached to the index.

lirsNextToken - The pagination token.

lirsResponseStatus - -- | The response status code.

data ListIndexResponse

See: listIndexResponse smart constructor.

Response Lenses

lirsIndexAttachments :: Lens' ListIndexResponse [IndexAttachment]

The objects and indexed values attached to the index.

lirsNextToken :: Lens' ListIndexResponse (Maybe Text)

The pagination token.

lirsResponseStatus :: Lens' ListIndexResponse Int

- | The response status code.