Network.AWS.CloudDirectory.ListObjectAttributes

Description

Lists all attributes that are associated with an object.

Creating a Request

listObjectAttributes

Arguments

:: Text
loaDirectoryARN

-> ObjectReference
loaObjectReference

-> ListObjectAttributes

Creates a value of ListObjectAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loaFacetFilter - Used to filter the list of object attributes that are associated with a certain facet.

loaConsistencyLevel - Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

loaNextToken - The pagination token.

loaMaxResults - The maximum number of items to be retrieved in a single call. This is an approximate number.

loaDirectoryARN - The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns .

loaObjectReference - The reference that identifies the object whose attributes will be listed.

data ListObjectAttributes

See: listObjectAttributes smart constructor.

Request Lenses

loaFacetFilter :: Lens' ListObjectAttributes (Maybe SchemaFacet)

Used to filter the list of object attributes that are associated with a certain facet.

loaConsistencyLevel :: Lens' ListObjectAttributes (Maybe ConsistencyLevel)

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

loaNextToken :: Lens' ListObjectAttributes (Maybe Text)

The pagination token.

loaMaxResults :: Lens' ListObjectAttributes (Maybe Natural)

The maximum number of items to be retrieved in a single call. This is an approximate number.

loaDirectoryARN :: Lens' ListObjectAttributes Text

The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns .

loaObjectReference :: Lens' ListObjectAttributes ObjectReference

The reference that identifies the object whose attributes will be listed.

Destructuring the Response

listObjectAttributesResponse

Arguments

:: Int
loarsResponseStatus

-> ListObjectAttributesResponse

Creates a value of ListObjectAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loarsNextToken - The pagination token.

loarsAttributes - Attributes map that is associated with the object. AttributeArn is the key, and attribute value is the value.

loarsResponseStatus - -- | The response status code.

data ListObjectAttributesResponse

See: listObjectAttributesResponse smart constructor.

Response Lenses

loarsNextToken :: Lens' ListObjectAttributesResponse (Maybe Text)

The pagination token.

loarsAttributes :: Lens' ListObjectAttributesResponse [AttributeKeyAndValue]

Attributes map that is associated with the object. AttributeArn is the key, and attribute value is the value.

loarsResponseStatus :: Lens' ListObjectAttributesResponse Int

- | The response status code.