Network.AWS.CloudDirectory.ListObjectChildren

Description

Returns a paginated list of child objects that are associated with a given object.

Creating a Request

listObjectChildren

Arguments

:: Text
locDirectoryARN

-> ObjectReference
locObjectReference

-> ListObjectChildren

Creates a value of ListObjectChildren with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

locConsistencyLevel - Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

locNextToken - The pagination token.

locMaxResults - The maximum number of items to be retrieved in a single call. This is an approximate number.

locDirectoryARN - The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns .

locObjectReference - The reference that identifies the object for which child objects are being listed.

data ListObjectChildren

See: listObjectChildren smart constructor.

Request Lenses

locConsistencyLevel :: Lens' ListObjectChildren (Maybe ConsistencyLevel)

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

locNextToken :: Lens' ListObjectChildren (Maybe Text)

The pagination token.

locMaxResults :: Lens' ListObjectChildren (Maybe Natural)

The maximum number of items to be retrieved in a single call. This is an approximate number.

locDirectoryARN :: Lens' ListObjectChildren Text

The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns .

locObjectReference :: Lens' ListObjectChildren ObjectReference

The reference that identifies the object for which child objects are being listed.

Destructuring the Response

listObjectChildrenResponse

Arguments

:: Int
locrsResponseStatus

-> ListObjectChildrenResponse

Creates a value of ListObjectChildrenResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

locrsChildren - Children structure, which is a map with key as the LinkName and ObjectIdentifier as the value.

locrsNextToken - The pagination token.

locrsResponseStatus - -- | The response status code.

data ListObjectChildrenResponse

See: listObjectChildrenResponse smart constructor.

Response Lenses

locrsChildren :: Lens' ListObjectChildrenResponse (HashMap Text Text)

Children structure, which is a map with key as the LinkName and ObjectIdentifier as the value.

locrsNextToken :: Lens' ListObjectChildrenResponse (Maybe Text)

The pagination token.

locrsResponseStatus :: Lens' ListObjectChildrenResponse Int

- | The response status code.