Network.AWS.CloudDirectory.ListObjectParents

Description

Lists parent objects that are associated with a given object in pagination fashion.

Creating a Request

listObjectParents

Arguments

:: Text
lopDirectoryARN

-> ObjectReference
lopObjectReference

-> ListObjectParents

Creates a value of ListObjectParents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lopConsistencyLevel - Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

lopNextToken - The pagination token.

lopMaxResults - The maximum number of items to be retrieved in a single call. This is an approximate number.

lopDirectoryARN - The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns .

lopObjectReference - The reference that identifies the object for which parent objects are being listed.

data ListObjectParents

See: listObjectParents smart constructor.

Request Lenses

lopConsistencyLevel :: Lens' ListObjectParents (Maybe ConsistencyLevel)

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

lopNextToken :: Lens' ListObjectParents (Maybe Text)

The pagination token.

lopMaxResults :: Lens' ListObjectParents (Maybe Natural)

The maximum number of items to be retrieved in a single call. This is an approximate number.

lopDirectoryARN :: Lens' ListObjectParents Text

The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns .

lopObjectReference :: Lens' ListObjectParents ObjectReference

The reference that identifies the object for which parent objects are being listed.

Destructuring the Response

listObjectParentsResponse

Arguments

:: Int
lrsResponseStatus

-> ListObjectParentsResponse

Creates a value of ListObjectParentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrsNextToken - The pagination token.

lrsParents - The parent structure, which is a map with key as the ObjectIdentifier and LinkName as the value.

lrsResponseStatus - -- | The response status code.

data ListObjectParentsResponse

See: listObjectParentsResponse smart constructor.

Response Lenses

lrsNextToken :: Lens' ListObjectParentsResponse (Maybe Text)

The pagination token.

lrsParents :: Lens' ListObjectParentsResponse (HashMap Text Text)

The parent structure, which is a map with key as the ObjectIdentifier and LinkName as the value.

lrsResponseStatus :: Lens' ListObjectParentsResponse Int

- | The response status code.