Network.AWS.CloudDirectory.ListObjectPolicies

Description

Returns policies attached to an object in pagination fashion.

Creating a Request

listObjectPolicies

Arguments

:: Text
lDirectoryARN

-> ObjectReference
lObjectReference

-> ListObjectPolicies

Creates a value of ListObjectPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lConsistencyLevel - Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

lNextToken - The pagination token.

lMaxResults - The maximum number of items to be retrieved in a single call. This is an approximate number.

lDirectoryARN - The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns .

lObjectReference - Reference that identifies the object for which policies will be listed.

data ListObjectPolicies

See: listObjectPolicies smart constructor.

Request Lenses

lConsistencyLevel :: Lens' ListObjectPolicies (Maybe ConsistencyLevel)

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

lNextToken :: Lens' ListObjectPolicies (Maybe Text)

The pagination token.

lMaxResults :: Lens' ListObjectPolicies (Maybe Natural)

The maximum number of items to be retrieved in a single call. This is an approximate number.

lDirectoryARN :: Lens' ListObjectPolicies Text

The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns .

lObjectReference :: Lens' ListObjectPolicies ObjectReference

Reference that identifies the object for which policies will be listed.

Destructuring the Response

listObjectPoliciesResponse

Arguments

:: Int
loprsResponseStatus

-> ListObjectPoliciesResponse

Creates a value of ListObjectPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loprsNextToken - The pagination token.

loprsAttachedPolicyIds - A list of policy ObjectIdentifiers , that are attached to the object.

loprsResponseStatus - -- | The response status code.

data ListObjectPoliciesResponse

See: listObjectPoliciesResponse smart constructor.

Response Lenses

loprsNextToken :: Lens' ListObjectPoliciesResponse (Maybe Text)

The pagination token.

loprsAttachedPolicyIds :: Lens' ListObjectPoliciesResponse [Text]

A list of policy ObjectIdentifiers , that are attached to the object.

loprsResponseStatus :: Lens' ListObjectPoliciesResponse Int

- | The response status code.