Network.AWS.CloudDirectory.ListOutgoingTypedLinks

Description

Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed link .

Creating a Request

listOutgoingTypedLinks

Arguments

:: Text
lotlDirectoryARN

-> ObjectReference
lotlObjectReference

-> ListOutgoingTypedLinks

Creates a value of ListOutgoingTypedLinks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lotlFilterAttributeRanges - Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

lotlConsistencyLevel - The consistency level to execute the request at.

lotlNextToken - The pagination token.

lotlFilterTypedLink - Filters are interpreted in the order of the attributes defined on the typed link facet, not the order they are supplied to any API calls.

lotlMaxResults - The maximum number of results to retrieve.

lotlDirectoryARN - The Amazon Resource Name (ARN) of the directory where you want to list the typed links.

lotlObjectReference - A reference that identifies the object whose attributes will be listed.

data ListOutgoingTypedLinks

See: listOutgoingTypedLinks smart constructor.

Request Lenses

lotlFilterAttributeRanges :: Lens' ListOutgoingTypedLinks [TypedLinkAttributeRange]

Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

lotlConsistencyLevel :: Lens' ListOutgoingTypedLinks (Maybe ConsistencyLevel)

The consistency level to execute the request at.

lotlNextToken :: Lens' ListOutgoingTypedLinks (Maybe Text)

The pagination token.

lotlFilterTypedLink :: Lens' ListOutgoingTypedLinks (Maybe TypedLinkSchemaAndFacetName)

Filters are interpreted in the order of the attributes defined on the typed link facet, not the order they are supplied to any API calls.

lotlMaxResults :: Lens' ListOutgoingTypedLinks (Maybe Natural)

The maximum number of results to retrieve.

lotlDirectoryARN :: Lens' ListOutgoingTypedLinks Text

The Amazon Resource Name (ARN) of the directory where you want to list the typed links.

lotlObjectReference :: Lens' ListOutgoingTypedLinks ObjectReference

A reference that identifies the object whose attributes will be listed.

Destructuring the Response

listOutgoingTypedLinksResponse

Arguments

:: Int
lotlrsResponseStatus

-> ListOutgoingTypedLinksResponse

Creates a value of ListOutgoingTypedLinksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lotlrsTypedLinkSpecifiers - Returns a typed link specifier as output.

lotlrsNextToken - The pagination token.

lotlrsResponseStatus - -- | The response status code.

data ListOutgoingTypedLinksResponse

See: listOutgoingTypedLinksResponse smart constructor.

Response Lenses

lotlrsTypedLinkSpecifiers :: Lens' ListOutgoingTypedLinksResponse [TypedLinkSpecifier]

Returns a typed link specifier as output.

lotlrsNextToken :: Lens' ListOutgoingTypedLinksResponse (Maybe Text)

The pagination token.

lotlrsResponseStatus :: Lens' ListOutgoingTypedLinksResponse Int

- | The response status code.