Network.AWS.CloudDirectory.ListPolicyAttachments

Description

Returns all of the ObjectIdentifiers to which a given policy is attached.

Creating a Request

listPolicyAttachments

Arguments

:: Text
lpaDirectoryARN

-> ObjectReference
lpaPolicyReference

-> ListPolicyAttachments

Creates a value of ListPolicyAttachments with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpaConsistencyLevel - Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

lpaNextToken - The pagination token.

lpaMaxResults - The maximum number of items to be retrieved in a single call. This is an approximate number.

lpaDirectoryARN - The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns .

lpaPolicyReference - The reference that identifies the policy object.

data ListPolicyAttachments

See: listPolicyAttachments smart constructor.

Request Lenses

lpaConsistencyLevel :: Lens' ListPolicyAttachments (Maybe ConsistencyLevel)

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

lpaNextToken :: Lens' ListPolicyAttachments (Maybe Text)

The pagination token.

lpaMaxResults :: Lens' ListPolicyAttachments (Maybe Natural)

The maximum number of items to be retrieved in a single call. This is an approximate number.

lpaDirectoryARN :: Lens' ListPolicyAttachments Text

The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns .

lpaPolicyReference :: Lens' ListPolicyAttachments ObjectReference

The reference that identifies the policy object.

Destructuring the Response

listPolicyAttachmentsResponse

Arguments

:: Int
lparsResponseStatus

-> ListPolicyAttachmentsResponse

Creates a value of ListPolicyAttachmentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lparsObjectIdentifiers - A list of ObjectIdentifiers to which the policy is attached.

lparsNextToken - The pagination token.

lparsResponseStatus - -- | The response status code.

data ListPolicyAttachmentsResponse

See: listPolicyAttachmentsResponse smart constructor.

Response Lenses

lparsObjectIdentifiers :: Lens' ListPolicyAttachmentsResponse [Text]

A list of ObjectIdentifiers to which the policy is attached.

lparsNextToken :: Lens' ListPolicyAttachmentsResponse (Maybe Text)

The pagination token.

lparsResponseStatus :: Lens' ListPolicyAttachmentsResponse Int

- | The response status code.