Network.AWS.CloudDirectory.ListTagsForResource

Description

Returns tags for a resource. Tagging is currently supported only for directories with a limit of 50 tags per directory. All 50 tags are returned for a given directory with this API call.

Creating a Request

listTagsForResource

Arguments

:: Text
ltfrResourceARN

-> ListTagsForResource

Creates a value of ListTagsForResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltfrNextToken - The pagination token. This is for future use. Currently pagination is not supported for tagging.

ltfrMaxResults - The MaxResults parameter sets the maximum number of results returned in a single page. This is for future use and is not supported currently.

ltfrResourceARN - The Amazon Resource Name (ARN) of the resource. Tagging is only supported for directories.

data ListTagsForResource

See: listTagsForResource smart constructor.

Request Lenses

ltfrNextToken :: Lens' ListTagsForResource (Maybe Text)

The pagination token. This is for future use. Currently pagination is not supported for tagging.

ltfrMaxResults :: Lens' ListTagsForResource (Maybe Natural)

The MaxResults parameter sets the maximum number of results returned in a single page. This is for future use and is not supported currently.

ltfrResourceARN :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource. Tagging is only supported for directories.

Destructuring the Response

listTagsForResourceResponse

Arguments

:: Int
ltfrrsResponseStatus

-> ListTagsForResourceResponse

Creates a value of ListTagsForResourceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltfrrsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

ltfrrsTags - A list of tag key value pairs that are associated with the response.

ltfrrsResponseStatus - -- | The response status code.

data ListTagsForResourceResponse

See: listTagsForResourceResponse smart constructor.

Response Lenses

ltfrrsNextToken :: Lens' ListTagsForResourceResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

ltfrrsTags :: Lens' ListTagsForResourceResponse [Tag]

A list of tag key value pairs that are associated with the response.

ltfrrsResponseStatus :: Lens' ListTagsForResourceResponse Int

- | The response status code.