Network.AWS.CloudDirectory.ListTypedLinkFacetAttributes

Description

Returns a paginated list of all attribute definitions for a particular TypedLinkFacet . For more information, see Typed link .

Creating a Request

listTypedLinkFacetAttributes

Arguments

:: Text
ltlfaSchemaARN

-> Text
ltlfaName

-> ListTypedLinkFacetAttributes

Creates a value of ListTypedLinkFacetAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltlfaNextToken - The pagination token.

ltlfaMaxResults - The maximum number of results to retrieve.

ltlfaSchemaARN - The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns .

ltlfaName - The unique name of the typed link facet.

data ListTypedLinkFacetAttributes

See: listTypedLinkFacetAttributes smart constructor.

Request Lenses

ltlfaNextToken :: Lens' ListTypedLinkFacetAttributes (Maybe Text)

The pagination token.

ltlfaMaxResults :: Lens' ListTypedLinkFacetAttributes (Maybe Natural)

The maximum number of results to retrieve.

ltlfaSchemaARN :: Lens' ListTypedLinkFacetAttributes Text

The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns .

ltlfaName :: Lens' ListTypedLinkFacetAttributes Text

The unique name of the typed link facet.

Destructuring the Response

listTypedLinkFacetAttributesResponse

Arguments

:: Int
ltlfarsResponseStatus

-> ListTypedLinkFacetAttributesResponse

Creates a value of ListTypedLinkFacetAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltlfarsNextToken - The pagination token.

ltlfarsAttributes - An ordered set of attributes associate with the typed link.

ltlfarsResponseStatus - -- | The response status code.

data ListTypedLinkFacetAttributesResponse

See: listTypedLinkFacetAttributesResponse smart constructor.

Response Lenses

ltlfarsNextToken :: Lens' ListTypedLinkFacetAttributesResponse (Maybe Text)

The pagination token.

ltlfarsAttributes :: Lens' ListTypedLinkFacetAttributesResponse [TypedLinkAttributeDefinition]

An ordered set of attributes associate with the typed link.

ltlfarsResponseStatus :: Lens' ListTypedLinkFacetAttributesResponse Int

- | The response status code.