Network.AWS.CloudDirectory.ListTypedLinkFacetNames

Description

Returns a paginated list of TypedLink facet names for a particular schema. For more information, see Typed link .

Creating a Request

listTypedLinkFacetNames

Arguments

:: Text
ltlfnSchemaARN

-> ListTypedLinkFacetNames

Creates a value of ListTypedLinkFacetNames with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltlfnNextToken - The pagination token.

ltlfnMaxResults - The maximum number of results to retrieve.

ltlfnSchemaARN - The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns .

data ListTypedLinkFacetNames

See: listTypedLinkFacetNames smart constructor.

Request Lenses

ltlfnNextToken :: Lens' ListTypedLinkFacetNames (Maybe Text)

The pagination token.

ltlfnMaxResults :: Lens' ListTypedLinkFacetNames (Maybe Natural)

The maximum number of results to retrieve.

ltlfnSchemaARN :: Lens' ListTypedLinkFacetNames Text

The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns .

Destructuring the Response

listTypedLinkFacetNamesResponse

Arguments

:: Int
ltlfnrsResponseStatus

-> ListTypedLinkFacetNamesResponse

Creates a value of ListTypedLinkFacetNamesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltlfnrsNextToken - The pagination token.

ltlfnrsFacetNames - The names of typed link facets that exist within the schema.

ltlfnrsResponseStatus - -- | The response status code.

data ListTypedLinkFacetNamesResponse

See: listTypedLinkFacetNamesResponse smart constructor.

Response Lenses

ltlfnrsNextToken :: Lens' ListTypedLinkFacetNamesResponse (Maybe Text)

The pagination token.

ltlfnrsFacetNames :: Lens' ListTypedLinkFacetNamesResponse [Text]

The names of typed link facets that exist within the schema.

ltlfnrsResponseStatus :: Lens' ListTypedLinkFacetNamesResponse Int

- | The response status code.