Network.AWS.CloudDirectory.LookupPolicy

Description

Lists all policies from the root of the Directory to the object specified. If there are no policies present, an empty list is returned. If policies are present, and if some objects don't have the policies attached, it returns the ObjectIdentifier for such objects. If policies are present, it returns ObjectIdentifier , policyId , and policyType . Paths that don't lead to the root from the target object are ignored. For more information, see Policies .

Creating a Request

lookupPolicy

Arguments

:: Text
lpDirectoryARN

-> ObjectReference
lpObjectReference

-> LookupPolicy

Creates a value of LookupPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpNextToken - The token to request the next page of results.

lpMaxResults - The maximum number of items to be retrieved in a single call. This is an approximate number.

lpDirectoryARN - The Amazon Resource Name (ARN) that is associated with the Directory . For more information, see arns .

lpObjectReference - Reference that identifies the object whose policies will be looked up.

data LookupPolicy

See: lookupPolicy smart constructor.

Request Lenses

lpNextToken :: Lens' LookupPolicy (Maybe Text)

The token to request the next page of results.

lpMaxResults :: Lens' LookupPolicy (Maybe Natural)

The maximum number of items to be retrieved in a single call. This is an approximate number.

lpDirectoryARN :: Lens' LookupPolicy Text

The Amazon Resource Name (ARN) that is associated with the Directory . For more information, see arns .

lpObjectReference :: Lens' LookupPolicy ObjectReference

Reference that identifies the object whose policies will be looked up.

Destructuring the Response

lookupPolicyResponse

Arguments

:: Int
lprsResponseStatus

-> LookupPolicyResponse

Creates a value of LookupPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lprsNextToken - The pagination token.

lprsPolicyToPathList - Provides list of path to policies. Policies contain PolicyId , ObjectIdentifier , and PolicyType . For more information, see Policies .

lprsResponseStatus - -- | The response status code.

data LookupPolicyResponse

See: lookupPolicyResponse smart constructor.

Response Lenses

lprsNextToken :: Lens' LookupPolicyResponse (Maybe Text)

The pagination token.

lprsPolicyToPathList :: Lens' LookupPolicyResponse [PolicyToPath]

Provides list of path to policies. Policies contain PolicyId , ObjectIdentifier , and PolicyType . For more information, see Policies .

lprsResponseStatus :: Lens' LookupPolicyResponse Int

- | The response status code.