Network.AWS.CloudDirectory.PutSchemaFromJSON

Description

Allows a schema to be updated using JSON upload. Only available for development schemas. See JSON Schema Format for more information.

Creating a Request

putSchemaFromJSON

Arguments

:: Text
psfjSchemaARN

-> Text
psfjDocument

-> PutSchemaFromJSON

Creates a value of PutSchemaFromJSON with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psfjSchemaARN - The ARN of the schema to update.

psfjDocument - The replacement JSON schema.

data PutSchemaFromJSON

See: putSchemaFromJSON smart constructor.

Request Lenses

psfjSchemaARN :: Lens' PutSchemaFromJSON Text

The ARN of the schema to update.

psfjDocument :: Lens' PutSchemaFromJSON Text

The replacement JSON schema.

Destructuring the Response

putSchemaFromJSONResponse

Arguments

:: Int
psfjrsResponseStatus

-> PutSchemaFromJSONResponse

Creates a value of PutSchemaFromJSONResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psfjrsARN - The ARN of the schema to update.

psfjrsResponseStatus - -- | The response status code.

data PutSchemaFromJSONResponse

See: putSchemaFromJSONResponse smart constructor.

Response Lenses

psfjrsARN :: Lens' PutSchemaFromJSONResponse (Maybe Text)

The ARN of the schema to update.

psfjrsResponseStatus :: Lens' PutSchemaFromJSONResponse Int

- | The response status code.