Network.AWS.CloudDirectory.UpdateTypedLinkFacet

Description

Updates a TypedLinkFacet . For more information, see Typed link .

Creating a Request

updateTypedLinkFacet

Arguments

:: Text
utlfSchemaARN

-> Text
utlfName

-> UpdateTypedLinkFacet

Creates a value of UpdateTypedLinkFacet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utlfSchemaARN - The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns .

utlfName - The unique name of the typed link facet.

utlfAttributeUpdates - Attributes update structure.

utlfIdentityAttributeOrder - The order of identity attributes for the facet, from most significant to least significant. The ability to filter typed links considers the order that the attributes are defined on the typed link facet. When providing ranges to a typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range. Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls. For more information about identity attributes, see Typed link .

data UpdateTypedLinkFacet

See: updateTypedLinkFacet smart constructor.

Request Lenses

utlfSchemaARN :: Lens' UpdateTypedLinkFacet Text

The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns .

utlfName :: Lens' UpdateTypedLinkFacet Text

The unique name of the typed link facet.

utlfAttributeUpdates :: Lens' UpdateTypedLinkFacet [TypedLinkFacetAttributeUpdate]

Attributes update structure.

utlfIdentityAttributeOrder :: Lens' UpdateTypedLinkFacet [Text]

The order of identity attributes for the facet, from most significant to least significant. The ability to filter typed links considers the order that the attributes are defined on the typed link facet. When providing ranges to a typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range. Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls. For more information about identity attributes, see Typed link .

Destructuring the Response

updateTypedLinkFacetResponse

Arguments

:: Int
utlfrsResponseStatus

-> UpdateTypedLinkFacetResponse

Creates a value of UpdateTypedLinkFacetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utlfrsResponseStatus - -- | The response status code.

data UpdateTypedLinkFacetResponse

See: updateTypedLinkFacetResponse smart constructor.

Response Lenses

utlfrsResponseStatus :: Lens' UpdateTypedLinkFacetResponse Int

- | The response status code.